ingest_worker/requirements.txt
paho-mqtt>=2.1.0,<3.0.0
psycopg[binary]>=3.2.0,<4.0.0
